Confirmación eliminar cuenta

+++ frontend/src/pages/user/Perfil.jsx
@@ -11,7 +11,7 @@ import UserTable from "../../components/UserTable";
 import { useState, useEffect } from 'react';
 import InputGroup from 'react-bootstrap/InputGroup';
 import MascotasTable from "../../components/MascotasTable";
-import { popError, popSuccess2 } from '../../utils/popUp';
+import { popError, popSuccess2, confirmDeletion } from '../../utils/popUp';
 
 const PanelAdmin = () => {
     return(
@@ -157,12 +157,14 @@ const Perfil = () => {
     
 
     const handleDelete = async () => {
-        try {
-          await axios.delete(`http://146.83.198.35:1273/api/usuarios/${auth._id}`);
-          popSuccess2('Eliminado Correctamente');
-        } catch (err) {
-          popError('No se ah logrado eliminar');
-          console.error('API Error:', err);
+        if (await confirmDeletion('¿Estás seguro de eliminar este usuario?')) {
+            try {
+              await axios.delete(`http://146.83.198.35:1273/api/usuarios/${auth._id}`);
+              popSuccess2('Eliminado Correctamente');
+            } catch (err) {
+              popError('No se ah logrado eliminar');
+              console.error('API Error:', err);
+            }
         }
       };
